feat: handle additional custom fields and epic
'customfield' is just added as a new column in the final result.
Note that it was only tested for one extra fields, it will probably
won't work with multiple.

For Epic, there is more processing: it create a couple of new column by joining the epic_id and epic_summary.
There is two possible epic meta info either in 'summary' or 'description'.
'summary' was choose as it is shorter and more readable than 'description'.